Move remaining stuff from sdmvalues to the symbolize tool's script file, since it was the only remaining reference. Delete sdmvalues.py.
